Fix fit(::Cauchy, x)
Due to an apparent change in the output of iqr(x), this fit method has
been broken for awhile and threw a BoundsError because it expected iqr()
to return two numbers.

This PR restores the functionality of fit(::Cauchy, x), by calling the
quantiles function directly. Doing so has the added benefit of getting
the median in the same computation. I've also generalized this fit
method to accept as input x::AbstractArray, not just x::Array.