A multivariate lognormal using MvNormal internally.
Implements the same constructors as MvNormal, as well as the interface
discussed in the Distributions documentation.

No changes to the existing code from Distributions.

The tests are very similar to the ones for MvNormal itself.

Also contains functionality to calculate the scale and location for a lognormal given
some desired statistics for the distribution.

This functionality is needed in e.g., MCMC methods, where you need to
center a distribution around e.g. the median or mode.

In particular, there are functions that allow to calculate:
(1) location and scale for a given mean and covariance
(2) location for a given scale and either mean, median or mode

(location and scale cannot be calculated analytically from e.g., mode and
covariance)

The added scale and location functions are the equivalent of static class
functions in C++ (typed on ::Type{MvLogNormal}) to ensure correct
dispatch.

I added tests to test/mvlognormal.jl for all functionality

Added documentation for the multivariate lognormal distribution.

Fixing indentation and removed AbstractMvLogNormal from exported types.