This fixes errors in the boundary conditions for univariate distributions.
These errors were partly caused because for some distributions
the boundary points of the support are not in the support itself.

I solved this by allowing the support to be a closed, half-open or open interval.
The default is a closed interval (which corresponds to the old behaviour), meaning that only
those distributions with a half-open or open support interval needed to be
ammended. The core of this functionality is in the file src/univariates.jl.
For half-open or open intervals, the macro @distr_boundaries specifies
which boundaries of the interval are open or closed. See e.g., in
src/univariate/continuous/lognormal.jl for an example.

Tests have been added to test/utils.jl and src/testutils.jl

A second set of errors came from sloppy implementations in pdf, logpdf,
cdf, ccdf, logcdf or logccdf functions of specific distributions
that did not explicitly test if the presented values fall inside or outside
the support. Tests that were added in src/testutils.jl flagged
- I believe - all instances of such errors, and the errors have been fixed.

I spotted 2 additional errors in the support() function in
src/univariates.jl.

There were no tests for this function so I added them to
src/testutils.jl/test_support()