fix infinite loop of Newton's method
quantile_newton and cquantile_newton find a quantile of a distribution
using Newton's method. However, their convergence will never be
satisfied when the initial point and the mode are the same. For example,
quantile_newton(Normal(0, 1), 0.5) will result in an infinite loop. This
patch fixes the problem by relaxing the convergence criterion.