Support larger VonMises concentrations (#818)
* Support larger VonMises concentrations

Previously, concentration parameters kappa > 700 led to AmosException, due to an
overflow when calling besseli(0, kappa).

The updated code uses besselix(.) instead of besseli(.), which avoids this
overflow for kappa > 700. The functions relying on the cached, and now
re-scaled, I0k are updated.

* Fix tolerance scaling sign in cdf of VonMises

As besselxi(j,k) = besseli(j,k) * exp(-k), the tolerance on besseli(.)
needs to be re-scaled by exp(-k) when applied to besselxi(.).

* Add test cases for VonMises with large concentration parameter

The target values for the test cases were determined by Mathematica.