WIP: Add support for the discrete normal distribution
The discrete normal distribution is defined as follows:
For an integer `x` in the support, the probability is proportional to

    P(x) = ℯ^(-(x-μ)^2/2σ^2)

The discrete normal distribution have significant applications in lattice
cryptography where noise parameters for integer lattices are generally drawn
from a multivariate discrete normal distribution. As such, some attention has
to be paid to the implementation of these samplers to make sure they are suitable
for cryptographic applications (disclaimer: I don't recommend doing so for production
cryptographic applications unless somebody audits these in much more detail).

There are four samplers for this distribution implemented in this PR:
 - Generic Sampling using an AliasTable (using the existing AliasTable sampler)
 - Generic Inversion sampling (using a new generic InversionCDT sampler)
 - Karney's original method [CFFK13]
 - KarneyFP (Karney's method with attention to floating point rounding) [MW19]

The implementations in this PR mostly follow the algorithms as presented in the
review article [MW19] (though the generic Karney is an implementation directly from
the original Karney paper).

The discrete Gaussian is represented as a type
```
struct DiscreteNormal{S<:Real, M<:Real} <: DiscreteUnivariateDistribution
    μ::M
    σ::S
    tail_cut::UnitRange{Int}
end
const DiscreteGaussian = DiscreteNormal
```

The mean and standard deviation may be any real type, not necessarily integer.
The type of σ determines the precision of the approximation of the distribution
to the true discrete Gaussian. The exact relationship depends on the sampler
used. See [MW19] for details.

If the tail cut is not explicitly provided, it is chosen depending on the
precision of σ such that the probability mass of any tail element is equal to
0.0 when properly rounded in the given floating point type (note that this does
not mean that the probability mass of the clipped tail is necessarily 0.0).
The tail cut need to be explicitly provided if σ is given in an arbitrary precision type.

There is a significant amount of literature on this topic and the samplers
in this PR don't even come close to covering it. If people are interested in
this topic [1] has additional references for other sampling algorithms.

[1] http://cseweb.ucsd.edu/~daniele/LatticeLinks/Sampling.html

[MW19] Michael Walter, "Sampling the Integers with Low Relative Error"
       (https://eprint.iacr.org/2019/068).

[CFFK13] Charles F. F. Karney "Sampling exactly from the normal distribution".
         https://arxiv.org/abs/1303.6257