Allow location to be >= 0 rather than strictly > 0.
This follows the Poisson distribution's @check_args, and,
admittedly, did not arise until I was pressing some extreme cases
of rare events and a sampler threw an exception due to location
being 0. (The location would never be in fact not zero, but perhaps
numerical roundoff might drive it there. -- I do not know the history
of Poisson distribution's @check_args, but I surmise something similar).