Improve type stability for truncated normal moments (#1717)
* Improve type stability for truncated normal moments

The internal functions used for computing the mean and variance of the
truncated normal distribution were implemented using literals like `√2`,
which are always `Float64`. However, some branches of the functions
based the return type off of the types of the arguments, which needn't
be `Float64`. This led to some instability that can be seen via
`@code_warntype`. For example, the inferred result of `_tnmom1` for
`Float32` inputs was `Union{Float32,Float64}`.

We can instead use the irrational constants defined for square roots of
pi and 2 and ratios thereof in order to avoid promoting to `Float64`
unnecessarily. This makes the return type concretely inferrable and
provides a small performance improvement, about 7 ns for `mean` and 20
ns for `var` on my machine as compared to current master.

* Apply suggestions from code review

Co-authored-by: David Widmann <devmotion@users.noreply.github.com>

---------

Co-authored-by: David Widmann <devmotion@users.noreply.github.com>