Handle truncation endpoints equal to support endpoints
Consider `truncated(Normal(), -Inf, Inf)`. Currently, this returns a
`Truncated`-wrapped `Normal`. However, truncating a normal distribution
to ±∞ is... not actually truncating it at all. We handle this case when
the bounds are specified as `nothing` and return the untruncated
distribution, but not when one or both bounds are equal to or beyond
the respective bounds of the support. If we recode `lower <= minimum(d)`
and/or `upper >= maximum(d)` to `nothing`, we're able to avoid a
`Truncated` wrapper in more cases. And when we do make a `Truncated`,
equivalent truncations can now compare equal and hit the same code paths
for dispatch on the truncation using `nothing` as appropriate.

Two things to note with this implementation:
- `truncated(d::D, a, b)` is no longer type stable since it can now
  return either `D` or `Truncated{D}` whereas before it only returned
  `D` if `a === b === nothing`.
- Nonsensical truncation, e.g. `truncated(Poisson(), -10, -1)`, still
  works and is still nonsensical.