Dirac: Implement location, scale, affine transformation
closes #1731

* location implemented
* scale implemented (= one)
* shifting = :+ implemented

* Scaling = :*, errors

Widening the Dirac cannot be done in a type stable way.
Dirac can be interpreted as limit of some distributions
in the limit σ^2 -> 0.0 (e.g. Normal, Uniform, Cosine).
Falling back to one of these is arbitrarily and not
type stable. Therefore it's up to the user what to do
when recaling (= lowering certainty) needs to be done.

Adding two Dirac distributions is allowed. The sum
of Dirac distributed random variables is interpreted
as the sum of their values. This result can be obtained
by either interpretation as certain bare numbers or
by convolution.

resolves #1695

Type T is reduced to T<:Real to be consistent with other
distributions. Non-scalar values are no longer allowed
as arguments.
This is a **potentially breaking** change, if user code
relied on unintended behaviour.