Define fallback implementations for mean, var, and entropy
We have the internal `expectation` function that uses `quadgk` to
compute integrals, as well as a fallback implementation for
`kldivergence` that uses `expectation`, so it seems reasonable to
similarly define fallbacks for other quantities trivially computable
using `expectation`: `mean`, `var`, and `entropy`. We could probably do
`skewness` and `kurtosis` as well.

NOTE: Skipping CI until I've actually done more with this since we run a
lot of CI on this repo

[ci skip]